Add files via upload
I managed to predict and plot the survival chances of passengers from Titanic. The model used logistic regression as I found it to be more precise than the Tree and Random Forest methods. The model isn't that good but for the most part I can say it did really well and the plot shows logical conclusions that can also be interpreted by humans themselves. But who wants to do it when we have computers for such tasks, am I right?